# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "LOGISTIK UNIKOM API",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000"
        }
    ],
    "paths": {
...
        "/api/pengumuman": {
            "get": {
                "tags": [
                    "Pengumuman"
                ],
                "summary": "List pengumuman",
                "operationId": "e18e6d8c9c384d2f98e051d1b9d8b33c",
                "responses": {
                    "200": {
                        "description": "List pengumuman berhasil diambil"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Pengumuman"
                ],
                "summary": "Tambah pengumuman",
                "operationId": "ef1a1de3985c3fdc64a7e61515a4ad27",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "judul"
                                ],
                                "properties": {
                                    "judul": {
                                        "type": "string",
                                        "example": "Pengumuman Libur Nasional"
                                    },
                                    "deskripsi": {
                                        "type": "string",
                                        "example": "Libur tanggal 17 Agustus"
                                    },
                                    "gambar": {
                                        "description": "Upload gambar pengumuman",
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Pengumuman berhasil dibuat"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/pengumuman/{id}": {
            "get": {
                "tags": [
                    "Pengumuman"
                ],
                "summary": "Detail pengumuman",
                "operationId": "ff9d68e9a77b7f38f0da55aba813ca4a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID pengumuman",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detail pengumuman berhasil diambil"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Pengumuman"
                ],
                "summary": "Update pengumuman",
                "description": "Gunakan POST + multipart/form-data jika update dengan gambar",
                "operationId": "7280226aee65d7a6dae5fa09788948f2",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "judul": {
                                        "type": "string",
                                        "example": "Update Pengumuman"
                                    },
                                    "deskripsi": {
                                        "type": "string",
                                        "example": "Isi pengumuman terbaru"
                                    },
                                    "gambar": {
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Pengumuman berhasil diupdate"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Pengumuman"
                ],
                "summary": "Hapus pengumuman",
                "operationId": "3c18d6a818f39a57e5f83f300300b8e2",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Pengumuman berhasil dihapus"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    },
...
}